Composite fixes from the wider fork network
Fixes gathered by scanning all 145 forks of bbogush/nand_programmer and
the open PRs on upstream and on every fork that had any. This base
already carries liqili's own work, so what is left is what the other
forks had and what nobody had.

Chip database: audit every row cycle count
------------------------------------------
CLAUDE.md listed this as unaudited, with only S34ML01G1 verified. Row
cycles is a literal count of address bytes clocked out, so a wrong value
sends an extra or missing address cycle on every read, program and erase.

Checked all 41 rows against ceil(ceil(log2(total_size/page_size))/8) and
corrected six 1 Gbit 2 KB-page parts from 3 to 2: K9F1G08U0D,
H27U1G8F2B, H27U1G8F2C, TC58NVG0S3ETA00, MT29F1G08ABAEA and W29N01GV.
K9F1G08U0E and S34ML01G2 were corrected the same way in the two commits
carried over from the previous branch.

The table corroborated itself before any change: seven of the thirteen
1 Gbit entries already said 2 -- including W29N01HVSINA while its own
sibling W29N01GV said 3 -- and all ten 2 Gbit entries said 3. Extending
liqili's S34ML01G1 finding (8463c4a) to the rest of the table.

H27UBG8T2A is left at 4 deliberately and documented: its total size is
4294967295, a clamped 2^32-1 rather than the real 4 GiB, so the derived
page count is wrong by construction and cannot settle the question.

New chip, from coflery/nand_programmer
--------------------------------------
F59L1G81MB (ESMT, ID C8 D1 80 95 40), commit f6aa2e7. Its sibling
F59L1G81A is already here, so only this row was missing; row cycles set
to 2 to match, where the fork shipped 3.

Qt host, from 7134956/nand_programmer (Andrei Paulau)
-----------------------------------------------------
Writer::handleWriteAck() failed the whole transfer when the ack arrived
split across two reads. It now returns 0 to ask for more data, as
handleBadBlock() and handleError() already did. Commit 060084f.

Firmware, from banoz liqili#5 (closed unmerged)
-----------------------------------------------
nand_programmer.h uses uint8_t and uint32_t but never included stdint.h.
It compiles today only because every translation unit that includes it
happens to pull stdint in first.

The same PR removed CHIP=STM32F10X_MD from the SPL build, which is how I
noticed the library was compiled for medium density while the app is
built -DSTM32F10X_HD. Now aligned. All 30 SPL sources compile
byte-identically under both densities, so this was a trap for whatever
gets added next rather than a live defect.

Firmware, not found in any fork
-------------------------------
Three 32-bit multiplies feeding 64-bit addresses overflow above 4 GB:
block * block_size in np_nand_erase()'s caller, page * page_size in
np_nand_erase(), and page * page_size in np_send_bad_blocks().

Qt host, from upstream PRs #99 (7134956) and #103 (biaoge88v)
-------------------------------------------------------------
Both propose the same thing: confirm before writing a chip. Taken in
#103's shape -- validate the selection first so the prompt can name the
chip, parent the dialog, default to Cancel -- but not its diff, which is
based on an older tree. Placed before the file is opened and ahead of the
existing image-layout probe, which can still refuse the write after it.

Tests
-----
test_bad_block.c gains test_bitmap_indexing: marks every even block
across the full 8192-block table and requires every odd one to stay
clear, crossing all 1024 byte boundaries in both roles, then fills the
table. Covers the bit-index and shift errors the existing cases do not.
Replaces a parallel suite I had written against the old base, which
duplicated what this file already tested better.

Verified: firmware builds clean from firmware/ (text 60168, bss 21840),
Qt app builds clean under -Wextra -Werror via qt.pro, and the host suite
passes -- 11 NAND, 15 NOR, bad block and ECC, all 0 failures. cmake is
not installed here so the CMake path was not exercised.

+++ firmware/programmer/nand_programmer.c
@@ -351,7 +351,7 @@ static int np_read_bad_block_info_from_page(np_prog_t *prog, uint32_t block,
     uint32_t page, bool *is_bad)
 {
     uint32_t status;
-    uint64_t addr = block * prog->chip_info.block_size;
+    uint64_t addr = (uint64_t)block * prog->chip_info.block_size;
     uint8_t *bb_mark = &prog->page.buf[prog->chip_info.page_size +
         prog->chip_info.bb_mark_off];
 
@@ -472,7 +472,7 @@ static int np_nand_mark_bad(np_prog_t *prog, uint32_t page)
 static int np_nand_erase(np_prog_t *prog, uint32_t page)
 {
     uint32_t status;
-    uint64_t addr = page * prog->chip_info.page_size;
+    uint64_t addr = (uint64_t)page * prog->chip_info.page_size;
     
     DEBUG_PRINT("NAND erase at 0x%" PRIx64 "\r\n", addr);
 
@@ -1196,7 +1196,7 @@ static int np_send_bad_blocks(np_prog_t *prog)
 
     while (nand_bad_block_table_iter(&bb_iter, &page))
     {
-        if (np_send_bad_block_info(page * prog->chip_info.page_size,
+        if (np_send_bad_block_info((uint64_t)page * prog->chip_info.page_size,
             prog->chip_info.block_size, false))
         {
             return -1;
